improvements to correction calculation
added Decimal Format -> readability
fixed correction formula again
allowed 0.1 steps for real Time  and therefore changed datatype to double

+++ src/main/java/View/I_V_frame.java
@@ -185,7 +185,7 @@ public interface I_V_frame extends I_V_basic{
     /**
      * this method returns the value of the real Duration entered by the user
      */
-    int getRealTime();
+    double getRealTime();
 
     /**
      * this method updates all values of the correction factor panel
